fix(auth): durable override cleanup in the /login sweep, and the sign-in flow in its own modules
